@jfontan 2019-07-01
It seems that the biggest improvement are the changes in `worktree.go` and `storage/filesystem/index.go`. Can yo split this PR in two? One with changes for those files and another with changes for packfile code.

I've been reviewing the changes for packfile and make sense but in some cases they make the code slower. I'm pretty sure this can be improved to get good gains in both memory and speed when it's generating the index.

@mcuadros what do you think?