@jfontan 2019-07-18
@orisano I would love to take a look to the changes for packfile you did. Even if they are not increasing speed now they look very promising. Can you open a new PR with only the changes from the files in `plumbing/format/packfile/*`. We can work from there to improve them.

Thanks!