@jfontan 2019-07-31
The changes from this PR are now in #1180